extract general PropertyDataFetcherImpl which is independent of DataFetcherEnvironment
